Extract common exporter form logic to a common function
