+++ src/components/app/exporters/GoogleSheetsExporter.vue
@@ -39,35 +39,19 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { computed, ref, reactive } from '@vue/composition-api';
-import store from '@/store';
-import { GoogleSheetsConfig } from '@/originalBudgetTrackingApp/configManager/configManager';
-import { VForm } from '@/types/vuetify';
+import { OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
+import { setupExporterConfigForm } from '@/components/app/exporters/exportersCommon';
 
 export default Vue.extend({
   name: 'GoogleSheetsExporter',
 
   setup() {
-    const vForm = ref<VForm>();
-
-    const exporterName = 'googleSheets';
-    const exporter = reactive(JSON.parse(JSON.stringify(store.getters.Config.getExporter(exporterName))) as GoogleSheetsConfig);
-
-    const validated = ref(true);
-    const changed = ref(false);
-    const readyToSave = computed(() => validated && changed);
-    const submit = async () => {
-      if (vForm.value?.validate()) {
-        await store.dispatch.Config.updateExporter({ name: exporterName, exporter });
-        changed.value = false;
-      }
-    };
-
     return {
-      validated, changed, readyToSave, submit, exporter, vForm
+      ...setupExporterConfigForm(OutputVendorName.GOOGLE_SHEETS)
     };
   }
 });
+
 </script>
 
 <style scoped>
