+++ src/components/app/exporters/YnabExporter.vue
@@ -27,33 +27,15 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { computed, ref, reactive } from '@vue/composition-api';
-import store from '@/store';
-import { OutputVendorsConfig } from '@/originalBudgetTrackingApp/configManager/configManager';
-import { VForm } from '@/types/vuetify';
-
-const exporterName = 'ynab';
+import { setupExporterConfigForm } from '@/components/app/exporters/exportersCommon';
+import { OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
 
 export default Vue.extend({
   name: 'YnabExporter',
 
   setup() {
-    const vForm = ref<VForm>();
-
-    const exporter = reactive(JSON.parse(JSON.stringify(store.getters.Config.getExporter(exporterName))) as OutputVendorsConfig<typeof exporterName>);
-
-    const validated = ref(true);
-    const changed = ref(false);
-    const readyToSave = computed(() => validated && changed);
-    const submit = () => {
-      if (vForm.value?.validate()) {
-        store.dispatch.Config.updateExporter({ name: exporterName, exporter })
-          .then(() => { changed.value = false; });
-      }
-    };
-
     return {
-      validated, changed, readyToSave, submit, exporter, vForm
+      ...setupExporterConfigForm(OutputVendorName.YNAB)
     };
   }
 });
